chore: address PR feedback
- Updated release status endpoint in E2E tests
- Moved release logic to its own service
- Made internal types private
- Reused shared date functions
- used dayjs

RISDEV-7186
