+++ frontend/src/services/zielnormExpressionsService.ts
@@ -20,48 +20,6 @@ type ZielnormPreviewResponse = {
   expressions: ZielnormPreviewExpressionResponse[]
 }
 
-export type ZielnormReleaseStatus = {
-  normWorkEli: string
-  title: string
-  shortTitle: string
-  expressions: {
-    normExpressionEli: string
-    isGegenstandslos: boolean
-    currentStatus:
-      | "NOT_RELEASED"
-      | "PRAETEXT_RELEASED"
-      | "VOLLDOKUMENTATION_RELEASED"
-  }[]
-}
-export type ZielnormReleaseStatusDomain = {
-  normWorkEli: NormWorkEli
-  title: string
-  shortTitle: string
-  expressions: {
-    normExpressionEli: NormExpressionEli
-    isGegenstandslos: boolean
-    currentStatus:
-      | "NOT_RELEASED"
-      | "PRAETEXT_RELEASED"
-      | "VOLLDOKUMENTATION_RELEASED"
-  }[]
-}
-
-export function mapReleaseStatusResponseToDomain(
-  response: ZielnormReleaseStatus,
-): ZielnormReleaseStatusDomain {
-  return {
-    normWorkEli: NormWorkEli.fromString(response.normWorkEli),
-    title: response.title,
-    shortTitle: response.shortTitle,
-    expressions: response.expressions.map((expr) => ({
-      normExpressionEli: NormExpressionEli.fromString(expr.normExpressionEli),
-      isGegenstandslos: expr.isGegenstandslos,
-      currentStatus: expr.currentStatus,
-    })),
-  }
-}
-
 function mapResponseToDomain(
   response: ZielnormPreviewResponse,
 ): ZielnormPreview {
@@ -89,46 +47,6 @@ type useCreateZielnormExpressionsReturn = Omit<
   UseFetchReturn<ZielnormPreview>,
   "get" | "post" | "put" | "delete" | "patch" | "head" | "options"
 >
-
-type UseGetZielnormReleaseStatusReturn = Omit<
-  UseFetchReturn<ZielnormReleaseStatusDomain | null>,
-  "get" | "post" | "put" | "delete" | "patch" | "head" | "options"
->
-
-/**
- * Fetches the release statuses of expressions belonging to a Zielnorm.
- *
- * @param eli Work ELI of the Zielnorm
- * @param [fetchOptions={}] Additional options for the fetching
- * * @returns Reactive fetch wrapper for Zielnorm release status
- */
-export function useGetZielnormReleaseStatus(
-  eli: MaybeRefOrGetter<NormWorkEli | undefined>,
-  fetchOptions: UseFetchOptions = {},
-): UseGetZielnormReleaseStatusReturn {
-  const url = computed(() => {
-    const eliVal = toValue(eli)
-    if (!eliVal) return INVALID_URL
-    return `/${eliVal}/expressions/releasestatus`
-  })
-
-  const { data, ...rest } = useApiFetch(url, {
-    refetch: true,
-    ...fetchOptions,
-  }).json<ZielnormReleaseStatus>()
-
-  const mappedData = ref<ZielnormReleaseStatusDomain | null>(null)
-  watch(
-    data,
-    (newData) => {
-      if (newData === null) mappedData.value = null
-      else mappedData.value = mapReleaseStatusResponseToDomain(newData)
-    },
-    { immediate: true },
-  )
-
-  return { ...rest, data: mappedData }
-}
 
 /**
  * Fetches the list of Zielnorm previews from the API.
